Bug fix
There is an was an error in the console. It looks like this script should only be run on the page where the game is played. I removed the script from the page layout template and added it to the end of the rest-of-site template.
Everything still works, and now there are no erros in the console.